requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-ckeditor==6.4.0
django-gm2m==1.1.1
django-js-asset==2.0.0
django-modeltranslation==0.17.6
djangorestframework==3.13.1
Pillow==9.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
